the final task of the 14th sprint v.2

+++ src/main/java/ru/practicum/shareit/user/service/impl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.practicum.shareit.user.service.impl;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.exception.NotValidException;
 import ru.practicum.shareit.user.dto.UserDto;
@@ -9,8 +10,8 @@ import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.user.service.UserService;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 public class UserServiceImpl implements UserService {
 
@@ -22,33 +23,47 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto save(UserDto userDto) {
+        log.debug("Создание пользователя: {}", userDto);
         User user = UserMapper.toUser(userDto);
         if (userRepository.findByEmail(user.getEmail()).isPresent()) {
-            throw new NotValidException("Пользователь с такой электронной почтой уже существует.");
+            log.warn("Попытка создания пользователя с уже существующей электронной почтой: {}", user.getEmail());
+            throw new NotValidException("Пользователь с такой электронной почтой уже существует");
         }
         userRepository.save(user);
-        return UserMapper.toUserDto(user);
+        UserDto savedUser = UserMapper.toUserDto(user);
+        log.debug("Пользователь успешно создан: {}", savedUser);
+        return savedUser;
     }
 
     @Override
     public List<UserDto> getAll() {
-        return ((List<User>) userRepository.findAll())
+        log.debug("Получение информации о всех пользователях");
+        List<UserDto> foundUsers = userRepository.findAll()
                 .stream()
                 .map(UserMapper::toUserDto)
-                .collect(Collectors.toList());
+                .toList();
+        log.debug("Список найденных пользователей: {}", foundUsers);
+        return foundUsers;
     }
 
     @Override
     public UserDto getById(Long id) {
-        return userRepository.findById(id)
+        log.debug("Получение информации о пользователе по id: " + id);
+        UserDto foundUser = userRepository.findById(id)
                 .map(UserMapper::toUserDto)
                 .orElseThrow(() -> new IllegalArgumentException("Пользователь не найден"));
+        log.debug("По id {} найден пользователь: {}", id, foundUser);
+        return foundUser;
     }
 
     @Override
     public UserDto update(Long id, UserDto userDto) {
+        log.debug("Обновление данных пользователя по id: {}, новые данные: {}", id, userDto);
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Пользователь не найден"));
+                .orElseThrow(() -> {
+                    log.warn("Попытка обновления несуществующего пользователя с id: {}", id);
+                    return new IllegalArgumentException("Пользователь не найден");
+                });
 
         if (userDto.getName() != null) {
             user.setName(userDto.getName());
@@ -56,7 +71,8 @@ public class UserServiceImpl implements UserService {
 
         if (userDto.getEmail() != null) {
             if (userRepository.findByEmail(userDto.getEmail()).isPresent()) {
-                throw new NotValidException("Пользователь с такой электронной почтой уже существует.");
+                log.warn("Попытка создания пользователя с уже существующей электронной почтой: {}", user.getEmail());
+                throw new NotValidException("Пользователь с такой электронной почтой уже существует");
             }
             user.setEmail(userDto.getEmail());
         }
@@ -64,11 +80,19 @@ public class UserServiceImpl implements UserService {
         User userForUpdate = UserMapper.toUser(userDto);
 
         userRepository.update(id, userForUpdate);
-        return UserMapper.toUserDto(user);
+        UserDto updatedUser = UserMapper.toUserDto(user);
+        log.debug("Пользователь с id {} успешно обновлён: {}", id, updatedUser);
+        return updatedUser;
     }
 
     @Override
     public void delete(Long id) {
+        log.debug("Удаление пользователя с id: {}", id);
+        if (userRepository.findById(id).isEmpty()) {
+            log.warn("Попытка удаления несуществующего пользователя с id: {}", id);
+            throw new IllegalArgumentException("Пользователь не найден");
+        }
         userRepository.deleteById(id);
+        log.debug("Пользователь с id {} успешно удалён", id);
     }
 }
